requirements.txt
umap-learn
torch
open_clip_torch
tqdm
pandas
matplotlib